+++ tests/testthat/test-curl.R
@@ -134,3 +134,23 @@ test_that("can evaluate simple calls", {
   body <- resp_body_json(resp)
   expect_true(body$authenticated)
 })
+
+test_that("can read from clipboard", {
+  # need to skip on CI as can't read from clipboard there on Linux
+  skip_on_ci()
+  skip_if_not_installed("clipr")
+  # need to set env var so that `read/write_clip()` works in non-interactive mode
+  withr::local_envvar(CLIPR_ALLOW = TRUE)
+
+  # suppress warning because the clipboard might contain no readable text
+  old_clip <- suppressWarnings(clipr::read_clip())
+  withr::defer(clipr::write_clip(old_clip))
+  rlang::local_interactive()
+
+  clipr::write_clip("curl 'http://example.com' \\\n -H 'A: 1' \\\n -H 'B: 2'")
+  expect_snapshot({
+    curl_translate()
+    # also writes to clip
+    clipr::read_clip()
+  })
+})